requirements.txt
ipykernel==6.29.5
ipython==8.30.0
pandas==2.2.3
geopandas==1.0.1
polyline==2.0.2
python-dotenv==1.0.1
requests==2.32.3
shapely==2.0.6
